requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
pandas
openpyxl
python-multipart